refactor(net): split client utils tests for GH727
