+++ specs/GH727/tasks.md
@@ -210,24 +210,30 @@ GH-727 / #727
 - [x] `SP727-T197` Owner: coordinator. Done when: `src/utils/config/helpers.rs` keeps production config helper definitions and delegates tests with `#[path = "helpers_tests.rs"] mod tests;`. Verify: `tail -n 20 src/utils/config/helpers.rs`.
 - [x] `SP727-T198` Owner: coordinator. Done when: original inline config helper tests move to `src/utils/config/helpers_tests.rs` without assertion changes. Verify: `rg -n "test_env_utils_get_or_default|test_env_utils_get_as_bool|test_file_exists|test_find_config_file|test_validate_url|test_validate_duration_string_days|test_read_file|test_parse_yaml_file|test_write_yaml_file" src/utils/config/helpers_tests.rs`.
 - [x] `SP727-T199` Owner: verification owner. Done when: both touched config helper files are below U-16's 800-line ceiling and focused config helper tests pass. Verify: `wc -l src/utils/config/helpers.rs src/utils/config/helpers_tests.rs`; `cargo test utils::config::helpers --lib --all-features`.
-- [ ] `SP727-T200` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the config helpers tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
+- [x] `SP727-T200` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the config helpers tranche. Verify: #858 PR body, green PR CI, GraphQL reviewThreads totalCount 0, merge commit `8078679f965a`, #727 reopened, and #727 update comment.
+- [x] `SP727-T201` Owner: coordinator. Done when: after the config helpers tranche merges, the next #727 tranche is selected from the remaining queue with fresh tracked-file line-count evidence. Verify: at `origin/main@8078679f`, `src/utils/net/client/utils.rs` is 846 lines and 18 tracked Rust files remain over the U-16 ceiling.
+- [x] `SP727-T202` Owner: coordinator. Done when: the net client utils tranche documents why this slice is a test extraction rather than a production helper split. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T203` Owner: coordinator. Done when: `src/utils/net/client/utils.rs` keeps production net client helper definitions and delegates tests with `#[path = "utils_tests.rs"] mod tests;`. Verify: `tail -n 20 src/utils/net/client/utils.rs`.
+- [x] `SP727-T204` Owner: coordinator. Done when: original inline net client helper tests move to `src/utils/net/client/utils_tests.rs` without assertion changes. Verify: `rg -n "test_should_retry_rate_limited|test_calculate_retry_delay_exponential_backoff|test_get_timeout_openai|test_supports_httpx_timeout_openai|test_user_agent_openai|test_add_path_basic|test_validate_url_https|test_parse_content_type_simple|test_default_headers_has_content_type|test_extract_retry_after_seconds|test_create_http_client_default_config|test_get_environment_proxies_empty" src/utils/net/client/utils_tests.rs`.
+- [x] `SP727-T205` Owner: verification owner. Done when: both touched net client helper files are below U-16's 800-line ceiling and focused net client helper tests pass. Verify: `wc -l src/utils/net/client/utils.rs src/utils/net/client/utils_tests.rs`; `cargo test utils::net::client::utils --lib --all-features`.
+- [ ] `SP727-T206` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the net client utils tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for the config helpers file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the net client utils file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/utils/config/helpers.rs`
-- `src/utils/config/helpers_tests.rs`
+- `src/utils/net/client/utils.rs`
+- `src/utils/net/client/utils_tests.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
 - `git diff --check`
-- `cargo test utils::config::helpers --lib --all-features`
+- `cargo test utils::net::client::utils --lib --all-features`
 - `cargo check --lib --all-features`
 - `cargo check --all-features --locked`
 - `cargo check`
@@ -235,5 +241,5 @@ Writable ownership for this lane:
 
 ## Handoff Notes
 
-This PR is the next #727 config helpers test-extraction tranche and should use `Refs #727`, not `Closes #727`.
+This PR is the next #727 net client utils test-extraction tranche and should use `Refs #727`, not `Closes #727`.
 The issue should remain open until the final scan shows no Rust files over the U-16 ceiling.
